requirements.txt
scikit-image>=0.22
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
streamlit>=1.35
